agent/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
matplotlib>=3.5.0
numpy>=1.20.0
